Separate commitment prover and verifier to modules
